Made some generic classes for the different implementations.
Removed the static members of `RandomList` since the class was being instantiated, and the constructor was modifying static fields. It now also receives an input seed.

`BaseCalendar` is the base class, and `LinkedListCalendar` and `NestedArrayCalendar` now extend this class. `ImpNestedArrayCalendar` was changed to extend `NestedArrayCalendar`, and the redundant code was removed. The calendars also now have an iterator which is abstract in the `BaseCalendar` class for getting the event items for the report printing. This abstracts the difference between walking the array and traversing the linked list.

`EventArray` and `EventLink` now extend a `BaseEvent` which allows for the classes to be more generic.

`TestingApp` has 2 main changes - the seed for the `random` is reset to its original (now static) state for each test (allowing the same random numbers to be observed each run). The various `*Test()` methods were simplified down to one, which takes in an interface (`CalendarTestData`) which sets the name, the randomizer method, and the instance of the `BaseCalendar` to test.